Fix a bunch of unintended casts.
In the process, I discovered that all of the monster bolt and cone
attacks were inadvertently swapping their damage and rate. :(

+++ lib/src/engine/core/combat.dart
@@ -195,7 +195,7 @@ class Hit {
     var damage = (_attack.damage * _damageScale + _damageBonus) * resistScale;
     var damageCents = (damage * 100).toInt();
 
-    var rolled = rng.triangleInt(damageCents, damageCents ~/ 2);
+    var rolled = rng.triangleInt(damageCents, damageCents ~/ 2).toDouble();
     rolled *= getArmorMultiplier(armor);
     return (rolled / 100).round();
   }
